requirements.txt
fastapi
uvicorn
sqlalchemy
openai
python-dotenv
PyPDF2
pandas
pydantic
python-jobspy
httpx
beautifulsoup4
pytest
